pyproject.toml
[project]
name = "dashboard"
version = "0.1.0"
description = "DashBord For proudact"
authors = [
    {name = "Elhussin", email = "hasin3112@gmail.com"},
]
dependencies = ["django>=5.2.8", "djangorestframework>=3.16.1", "django-filter>=25.2", "django-mssql-backend>=2.8.1", "dotenv>=0.9.9"]
requires-python = ">=3.14"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
